+++ components.d.ts
@@ -7,9 +7,13 @@ export {}
 /* prettier-ignore */
 declare module 'vue' {
   export interface GlobalComponents {
+    DownloadingList: typeof import('./src/components/DownloadComponents/DownloadingList.vue')['default']
     DownloadPane: typeof import('./src/components/DownloadPane.vue')['default']
+    EpisodePane: typeof import('./src/components/DownloadComponents/EpisodePane.vue')['default']
     MangaDirIndicator: typeof import('./src/components/WatermarkComponents/MangaDirIndicator.vue')['default']
+    MangaEpisodePane: typeof import('./src/components/DownloadComponents/MangaEpisodePane.vue')['default']
     NButton: typeof import('naive-ui')['NButton']
+    NDivider: typeof import('naive-ui')['NDivider']
     NDropdown: typeof import('naive-ui')['NDropdown']
     NInput: typeof import('naive-ui')['NInput']
     NMessageProvider: typeof import('naive-ui')['NMessageProvider']
@@ -23,8 +27,11 @@ declare module 'vue' {
     NSwitch: typeof import('naive-ui')['NSwitch']
     NTabPane: typeof import('naive-ui')['NTabPane']
     NTabs: typeof import('naive-ui')['NTabs']
+    NText: typeof import('naive-ui')['NText']
     NTooltip: typeof import('naive-ui')['NTooltip']
     RemoveProgress: typeof import('./src/components/WatermarkComponents/RemoveProgress.vue')['default']
+    SearchMangaPane: typeof import('./src/components/DownloadComponents/SearchPane.vue')['default']
+    SearchPane: typeof import('./src/components/DownloadComponents/SearchPane.vue')['default']
     WatermarkCropper: typeof import('./src/components/WatermarkComponents/WatermarkCropper.vue')['default']
     WatermarkPane: typeof import('./src/components/WatermarkPane.vue')['default']
   }
