+++ src/components/DownloadComponents/DownloadingList.vue
@@ -1,0 +1,15 @@
+<script setup lang="ts">
+
+</script>
+
+<template>
+  <div>
+    下载列表
+    <n-progress type="line" :percentage="60" indicator-placement="inside"/>
+    <n-progress type="line" :percentage="60" indicator-placement="inside"/>
+    <n-progress type="line" :percentage="60" indicator-placement="inside"/>
+    <n-progress type="line" :percentage="60" indicator-placement="inside"/>
+    <n-progress type="line" :percentage="60" indicator-placement="inside"/>
+    <n-progress type="line" :percentage="60" indicator-placement="inside"/>
+  </div>
+</template>
